fix(catalogue): remove some function to utils
