+++ src/catalogue/demos/drag.tsx
@@ -6,16 +6,15 @@ import { cloneDeep } from 'lodash';
 import shortid from 'shortid';
 
 import { CatalogueProps } from '../components/catalogue';
+import { ITreeNode, useTreeData } from '../useTreeData';
 import {
     appendNodeByKey,
     findNodeByKey,
     findParentNodeByKey,
-    ITreeNode,
     removeEditNode,
     removeNodeByKey,
     updateTreeNodeEdit,
-    useTreeData,
-} from '../useTreeData';
+} from '../utils';
 
 interface IData {
     edit?: boolean;
